requirements.txt
pandas==1.4.2
numpy==1.22.3
pymongo==3.12.3
python-dotenv==0.20.0
requests==2.27.1
beautifulsoup4==4.11.1
lxml==4.8.0
praw==7.7.0
psycopg2-binary==2.9.3
boto3==1.24.1
apache-airflow-providers-amazon==6.0.0
apache-airflow-providers-postgres==5.3.1
apache-airflow-providers-mongo==3.1.1
plotly==5.8.0
matplotlib==3.5.2
fastapi==0.78.0
uvicorn==0.17.6
aiohttp>=3.8.0
s3fs
